refactor: move out CalendarProgram to its own component
